introduce new Injection & Binding configurators - remove anonymous classes
